TVB-2663 Remove tvb.adapter dependency from model_operation.py
