+++ framework_tvb/tvb/adapters/simulator/range_parameters.py
@@ -27,59 +27,14 @@
 #   Frontiers in Neuroinformatics (7:10. doi: 10.3389/fninf.2013.00010)
 #
 #
-import json
-import numpy
 from collections import OrderedDict
-from tvb.basic.neotraits.api import HasTraits, NArray, Range
+
+from tvb.basic.neotraits.api import NArray, Range
+from tvb.core.entities.transient.range_parameter import RangeParameter
 from tvb.datatypes.connectivity import Connectivity
 from tvb.datatypes.surfaces import Surface
 from tvb.simulator.integrators import IntegratorStochastic
 from tvb.simulator.simulator import Simulator
-
-
-class RangeParameter(object):
-    KEY_FLOAT_LO = 'lo'
-    KEY_FLOAT_STEP = 'step'
-    KEY_FLOAT_HI = 'hi'
-
-    def __init__(self, name, type, range_definition, is_array=False, range_values=None):
-        self.name = name
-        self.type = type
-        self.is_array = is_array
-        # range_definition is a Range object when type=float and a FilterChain otherwise
-        self.range_definition = range_definition
-        self.range_values = range_values
-
-    def to_json(self):
-        if self.type is float:
-            return json.dumps([self.name, {self.KEY_FLOAT_LO: self.range_definition.lo,
-                                           self.KEY_FLOAT_STEP: self.range_definition.step,
-                                           self.KEY_FLOAT_HI: self.range_definition.hi}])
-
-        return json.dumps([self.name, self.range_definition])
-
-    @staticmethod
-    def from_json(range_json):
-        range_list = json.loads(range_json)
-        if isinstance(range_list[1], dict):
-            return RangeParameter(range_list[0], float,
-                                  Range(range_list[1][RangeParameter.KEY_FLOAT_LO],
-                                        range_list[1][RangeParameter.KEY_FLOAT_HI],
-                                        range_list[1][RangeParameter.KEY_FLOAT_STEP]))
-
-        return RangeParameter(range_list[0], HasTraits, range_list[1])
-
-    def get_range_values(self):
-        if self.type is float:
-            self.range_values = self.range_definition.to_array()
-            if self.is_array:
-                self.range_values = [numpy.array([value]) for value in self.range_values]
-        return self.range_values
-
-    def fill_from_default(self, range_param):
-        # type: (RangeParameter) -> None
-        self.type = range_param.type
-        self.is_array = range_param.is_array
 
 
 class SimulatorRangeParameters(object):
