+++ framework_tvb/tvb/core/entities/model/model_operation.py
@@ -36,18 +36,19 @@ Here we define entities for Operations and Algorithms.
 .. moduleauthor:: Yann Gordon <yann@tvb.invalid>
 """
 
-import json
 import datetime
-from sqlalchemy.orm import relationship, backref
+import json
+
 from sqlalchemy import Boolean, Integer, String, DateTime, Column, ForeignKey
+from sqlalchemy.orm import relationship, backref
 from tvb.basic.logger.builder import get_logger
-from tvb.adapters.simulator.range_parameter import RangeParameter
 from tvb.config import TVB_IMPORTER_CLASS, TVB_IMPORTER_MODULE
-from tvb.core.neotraits.db import Base
-from tvb.core.utils import string2date, generate_guid
 from tvb.core.entities.exportable import Exportable
 from tvb.core.entities.model.model_project import Project, User
+from tvb.core.entities.transient.range_parameter import RangeParameter
+from tvb.core.neotraits.db import Base
 from tvb.core.utils import string2bool, date2string, LESS_COMPLEX_TIME_FORMAT
+from tvb.core.utils import string2date, generate_guid
 
 LOG = get_logger(__name__)
