+++ framework_tvb/tvb/interfaces/web/controllers/simulator_controller.py
@@ -35,7 +35,7 @@ from tvb.adapters.exporters.export_manager import ExportManager
 from tvb.adapters.simulator.equation_forms import get_form_for_equation
 from tvb.adapters.simulator.model_forms import get_form_for_model
 from tvb.adapters.simulator.noise_forms import get_form_for_noise
-from tvb.adapters.simulator.range_parameter import SimulatorRangeParameters
+from tvb.adapters.simulator.range_parameters import SimulatorRangeParameters
 from tvb.adapters.simulator.simulator_adapter import SimulatorAdapterForm
 from tvb.adapters.simulator.simulator_fragments import *
 from tvb.adapters.simulator.monitor_forms import get_form_for_monitor
